OoTR Compliance Update: Ganon's Castle
This is relatively modest, mostly cleaning up Shadow Trial honestly (though it does correctly tie the use of Bombchus to the Bomb Bag and not to the consumable drops as well). This does update the trial clear conditions to match OoTR so trials required can be properly supported even though it's not currently a feature.