Foolish Region Expansion
This should make it a bit easier for bigger regions to roll foolish. The items that can only possibly gate six or fewer checks have some checking added that allows them to fall into foolish regions if they only gate junk items or other non-relevant sometimes items. This code is rather inelegant, and I fly close to the sun on some self-referential recursion, but in terms of functionality, I think this works and does improve the hint system non-trivially. I expect other improvements under the hood to later render this obsolete, but for now, it should make for slightly better hints on average.