Improve Goron Bomb Jump trick logic
This mostly fully reworks the Oceanside Spider House logic to allow for Goron Bomb Jump as an access method, and in general I robustly went through OSH and made sure each check was correct. I'm hard assuming that behind skull 1 is the upstairs one and behind skull 2 is the downstairs one; flip them if this is backwards (we need to move away from naming checks by numbering like this). I also added in the Goron Bomb Jump to enter/exit the observatory area from termina field, but this is fairly trivial.