TS: Fix some GF hint stuff (#269)
This fixes an issue in which the GF Mask could erroneously be labeled as foolish when it was not in fairysanity. It will currently be a bit overbroad in what it handles as it's built to be ER safe. I also fix a very minor logic issue with one of the fairies in Woodfall Temple that should only come up