Minor MQ logic tweaks (Deku/Dodongo) (#308)
* Minor MQ logic tweaks (Deku/Dodongo)

This is a follow-up on Broject's PR, taking the clearly important stuff addressed there that should definitely be in the logic that isn't. The total impact of this should be relatively small, but this definitely improves the logic. Further revision and some of the stylistic clean-up involved with the macros may come later.

* Typo fix

)