MQ: Initial Spirit MQ logic (#315)
* Initial Spirit MQ logic

This dungeon is not nearly as complex as vanilla Spirit but is the most complex dungeon in MQ anyway. This definitely needs a look-over, but I think I got it right?

* A few further fixes

This fixes a few issues with Spirit MQ's logic file. It also fixes a minor error in Ganon's MQ logic file.