Gerudo Fortress ER Prep
This sets up Gerudo Fortress to be fully robust for entrance randomization. It fixes some issues with the Kakariko Potion Shop for ER as well (that only would have come up with interior shuffle, not yet implemented) and removes Epona's dependency on access to LLR to ride her. Gerudo Fortress is actually a really complex zone so this is kinda a lot, but it's what is necessary to make it randomizable fully. A likely future trick is a certain diagonal jump to skip the kitchen which has a place to insert it (a connection that evaluates to "false", and the unintended PoH in child GF is included commented out just in case we want to add it to the shuffle in the future.