Gerudo Fortress ER Prep (#370)
* Gerudo Fortress ER Prep

This sets up Gerudo Fortress to be fully robust for entrance randomization. It fixes some issues with the Kakariko Potion Shop for ER as well (that only would have come up with interior shuffle, not yet implemented) and removes Epona's dependency on access to LLR to ride her. Gerudo Fortress is actually a really complex zone so this is kinda a lot, but it's what is necessary to make it randomizable fully. A likely future trick is a certain diagonal jump to skip the kitchen which has a place to insert it (a connection that evaluates to "false", and the unintended PoH in child GF is included commented out just in case we want to add it to the shuffle in the future.

* Fixing up a few issues

You can Hookshot the roaming Gerudo guards. Going from kitchen bottom up this sucks too much, but it's fair enough to include in a few other places (Break Room, some exterior stuff). Also, Kak back is now a separate region that accounts for how you might not be able to get back there with truly nothing (though it has a great many ways you can access it).

* Small Kak Back Fixes

This makes Potion Shop Back, Granny Shop, and Kak Generic Grotto correctly exit to Kak Back. It also notes that Kak Back can only actually exit as an adult so even if you get in the back door as a kid you get stuck.