Improve error messages for incompatible ROMs
The text of the Error thrown when an incompatible ROM is selected is shown to users on the ootmm website. Many people don't understand what 'bad file hash' means and end up needing to ask about it on the ootmm discord.

This commit addresses this by suggesting the user use a ROM that is compatible.

This code is untested. I don't even know if it compiles.