All MM Enemies and Boss macros added.
Adjusted some logic in ikana graveyard for the bad bats to include Din's Fire for the Night 1 grave chest.

Also adjusted Gerudo Fortress logic to expect arrow use for the Gerudo Fighters. It is not that bad and there are arrows in the pots, generally. Added Gerudo Fighters to OOT Enemies, too.